fix: retry cancelled refresh compute and resubmit refresh on settings change (#676)
* refactor: move refresh state onto Session and task body into refresh module

The refresh epoch and diagnostic publish lock are session state, not
server plumbing — Session is the bag of state the server hands around,
so they live there now, with bump_refresh_epoch owning the counter
semantics. The background refresh task body (compute/apply/warm-up/
republish and its epoch stale-checks) moves out of server.rs into a
dedicated refresh module, leaving server.rs focused on the LanguageServer
handlers. SNAPSHOT_CANCEL_RETRIES moves to session.rs so both consumers
import it from the snapshot's owning module. No behavior change.


* fix: retry cancelled refresh compute and resubmit refresh on settings change

Two ways a project refresh could be dropped permanently, leaving stale
search paths until the next env change:

- A document edit during the off-lock compute cancels its Salsa queries.
  Nothing bumps the epoch or resubmits in that case, so the refresh was
  lost. Retry with a fresh database clone, bounded by the same retry
  budget as snapshot reads, before giving up with a warning.

- A diagnostics/tagspec-only config change bumped the epoch to guard
  publish ordering but never queued a replacement task, so a superseded
  in-flight env refresh lost its apply. Every relevant settings change
  now submits a refresh; its republish step covers the open-document
  diagnostics push the removed manual loop performed.


---------
