+++ crates/djls-server/src/server.rs
@@ -1,14 +1,8 @@
 use std::future::Future;
 use std::panic::AssertUnwindSafe;
 use std::sync::Arc;
-use std::sync::atomic::AtomicU64;
-use std::sync::atomic::Ordering;
 
 use djls_project::Db as ProjectDb;
-use djls_project::apply_refresh;
-use djls_project::compute_refresh;
-use djls_project::project_template_files;
-use djls_semantic::Db as SemanticDb;
 use djls_source::FileKind;
 use tokio::sync::Mutex;
 use tokio::sync::oneshot;
@@ -22,17 +16,15 @@ use crate::ext::PositionEncodingExt;
 use crate::ext::UriExt;
 use crate::logging::LoggingGuard;
 use crate::queue::Queue;
+use crate::refresh;
+use crate::session::SNAPSHOT_CANCEL_RETRIES;
 use crate::session::Session;
 use crate::session::SessionSnapshot;
-
-const SNAPSHOT_CANCEL_RETRIES: usize = 2;
 
 pub(crate) struct DjangoLanguageServer {
     client: Client,
     session: Arc<Mutex<Session>>,
     queue: Queue,
-    refresh_epoch: Arc<AtomicU64>,
-    diagnostic_publish_lock: Arc<Mutex<()>>,
     logging: LoggingGuard,
 }
 
@@ -43,8 +35,6 @@ impl DjangoLanguageServer {
             client,
             session: Arc::new(Mutex::new(Session::default())),
             queue: Queue::new(),
-            refresh_epoch: Arc::new(AtomicU64::new(0)),
-            diagnostic_publish_lock: Arc::new(Mutex::new(())),
             logging,
         }
     }
@@ -106,18 +96,23 @@ impl DjangoLanguageServer {
         unreachable!("snapshot retry loop must return")
     }
 
-    fn next_refresh_epoch(&self) -> u64 {
-        self.refresh_epoch.fetch_add(1, Ordering::AcqRel) + 1
-    }
-
-    async fn submit_project_refresh(&self, epoch: u64, log_initialization: bool) {
+    /// Bump the session's refresh epoch and queue a refresh task for the new
+    /// epoch. The task body lives in [`crate::refresh`].
+    async fn submit_project_refresh(&self, log_initialization: bool) {
         let client = self.client.clone();
-        let refresh_epoch = Arc::clone(&self.refresh_epoch);
-        let diagnostic_publish_lock = Arc::clone(&self.diagnostic_publish_lock);
+        let (epoch, refresh_epoch, diagnostic_publish_lock) = self
+            .with_session(|session| {
+                (
+                    session.bump_refresh_epoch(),
+                    session.refresh_epoch(),
+                    session.diagnostic_publish_lock(),
+                )
+            })
+            .await;
 
         let rx = self
             .with_session_mut_task(move |session| async move {
-                run_project_refresh_task(
+                refresh::run_project_refresh(
                     session,
                     client,
                     refresh_epoch,
@@ -185,7 +180,8 @@ impl DjangoLanguageServer {
 
         let diagnostic_count = diagnostics.len();
         let lsp_uri_text = lsp_uri.to_string();
-        let _publish_guard = self.diagnostic_publish_lock.lock().await;
+        let publish_lock = self.with_session(Session::diagnostic_publish_lock).await;
+        let _publish_guard = publish_lock.lock().await;
         self.client
             .publish_diagnostics(lsp_uri, diagnostics, Some(document.version()))
             .await;
@@ -196,264 +192,6 @@ impl DjangoLanguageServer {
             lsp_uri_text
         );
     }
-}
-
-fn refresh_is_stale(refresh_epoch: &AtomicU64, epoch: u64) -> bool {
-    refresh_epoch.load(Ordering::Acquire) != epoch
-}
-
-async fn run_project_refresh_task(
-    session: Arc<Mutex<Session>>,
-    client: Client,
-    refresh_epoch: Arc<AtomicU64>,
-    diagnostic_publish_lock: Arc<Mutex<()>>,
-    epoch: u64,
-    log_initialization: bool,
-) -> anyhow::Result<()> {
-    let start = std::time::Instant::now();
-    let result = run_project_refresh_task_inner(
-        session,
-        client,
-        refresh_epoch,
-        diagnostic_publish_lock,
-        epoch,
-    )
-    .await;
-
-    if log_initialization {
-        tracing::info!("Server initialization completed in {:?}", start.elapsed());
-    } else if result.is_ok() {
-        tracing::info!("Environment refresh completed in {:?}", start.elapsed());
-    }
-
-    result
-}
-
-async fn run_project_refresh_task_inner(
-    session: Arc<Mutex<Session>>,
-    client: Client,
-    refresh_epoch: Arc<AtomicU64>,
-    diagnostic_publish_lock: Arc<Mutex<()>>,
-    epoch: u64,
-) -> anyhow::Result<()> {
-    if refresh_is_stale(&refresh_epoch, epoch) {
-        tracing::debug!(
-            epoch,
-            "Skipping stale project refresh before locking session"
-        );
-        return Ok(());
-    }
-
-    let Some(compute_db) = ({
-        let session_lock = session.lock().await;
-        if refresh_is_stale(&refresh_epoch, epoch) {
-            tracing::debug!(
-                epoch,
-                "Skipping stale project refresh after locking session"
-            );
-            return Ok(());
-        }
-
-        let db = session_lock.db();
-        db.project().map(|_| db.clone())
-    }) else {
-        tracing::info!("Task: No project configured, skipping initialization.");
-        return Ok(());
-    };
-
-    let refresh = tokio::task::spawn_blocking(move || {
-        salsa::Cancelled::catch(AssertUnwindSafe(|| compute_refresh(&compute_db)))
-    })
-    .await
-    .expect("project refresh compute task must not panic");
-
-    let Some(refresh) = (match refresh {
-        Ok(refresh) => refresh,
-        Err(cancelled) => {
-            tracing::debug!(
-                ?cancelled,
-                "Project refresh compute cancelled; newer inputs will re-run refresh"
-            );
-            return Ok(());
-        }
-    }) else {
-        return Ok(());
-    };
-
-    if refresh_is_stale(&refresh_epoch, epoch) {
-        tracing::debug!(epoch, "Skipping stale project refresh before apply");
-        return Ok(());
-    }
-
-    let (snapshot, documents) = {
-        let mut session_lock = session.lock().await;
-        if refresh_is_stale(&refresh_epoch, epoch) {
-            tracing::debug!(epoch, "Skipping stale project refresh after apply lock");
-            return Ok(());
-        }
-
-        let db = session_lock.db_mut();
-        if db.project().is_none() {
-            return Ok(());
-        }
-
-        let t = std::time::Instant::now();
-        apply_refresh(db, refresh);
-        tracing::info!("External data refresh completed in {:?}", t.elapsed());
-
-        if refresh_is_stale(&refresh_epoch, epoch) {
-            tracing::debug!(epoch, "Skipping stale project refresh after apply");
-            return Ok(());
-        }
-
-        (session_lock.snapshot(), session_lock.open_documents())
-    };
-
-    warm_project_queries(snapshot.clone(), Arc::clone(&refresh_epoch), epoch).await;
-    republish_snapshot_diagnostics(
-        client,
-        snapshot,
-        documents,
-        refresh_epoch,
-        diagnostic_publish_lock,
-        epoch,
-    )
-    .await;
-
-    Ok(())
-}
-
-async fn warm_project_queries(
-    snapshot: SessionSnapshot,
-    refresh_epoch: Arc<AtomicU64>,
-    epoch: u64,
-) {
-    let result = tokio::task::spawn_blocking(move || {
-        salsa::Cancelled::catch(AssertUnwindSafe(|| {
-            let db = snapshot.db();
-            let Some(project) = db.project() else {
-                return;
-            };
-
-            if refresh_is_stale(&refresh_epoch, epoch) {
-                return;
-            }
-            let _ = db.tag_specs();
-
-            if refresh_is_stale(&refresh_epoch, epoch) {
-                return;
-            }
-            let _ = db.template_dirs();
-
-            if refresh_is_stale(&refresh_epoch, epoch) {
-                return;
-            }
-            let _ = db.template_libraries();
-
-            if refresh_is_stale(&refresh_epoch, epoch) {
-                return;
-            }
-            let _ = project_template_files(db, project);
-        }))
-    })
-    .await
-    .expect("project warm-up task must not panic");
-
-    if let Err(cancelled) = result {
-        tracing::debug!(
-            ?cancelled,
-            "Project refresh warm-up cancelled; newer inputs will re-warm queries"
-        );
-    }
-}
-
-async fn republish_snapshot_diagnostics(
-    client: Client,
-    snapshot: SessionSnapshot,
-    documents: Vec<TextDocument>,
-    refresh_epoch: Arc<AtomicU64>,
-    diagnostic_publish_lock: Arc<Mutex<()>>,
-    epoch: u64,
-) {
-    if snapshot.client_info().supports_pull_diagnostics() {
-        tracing::debug!("Client supports pull diagnostics, skipping refresh diagnostics push");
-        return;
-    }
-
-    for document in documents {
-        if refresh_is_stale(&refresh_epoch, epoch) {
-            tracing::debug!(epoch, "Skipping stale refresh diagnostics republish");
-            return;
-        }
-
-        let file = document.file();
-        let Some(diagnostics) = collect_snapshot_diagnostics(snapshot.clone(), file).await else {
-            continue;
-        };
-
-        if refresh_is_stale(&refresh_epoch, epoch) {
-            tracing::debug!(epoch, "Skipping stale refresh diagnostics publish");
-            return;
-        }
-
-        let Some(lsp_uri) = ls_types::Uri::from_path(document.path()) else {
-            continue;
-        };
-
-        let diagnostic_count = diagnostics.len();
-        let lsp_uri_text = lsp_uri.to_string();
-        let _publish_guard = diagnostic_publish_lock.lock().await;
-        if refresh_is_stale(&refresh_epoch, epoch) {
-            tracing::debug!(epoch, "Skipping stale refresh diagnostics publish");
-            return;
-        }
-        client
-            .publish_diagnostics(lsp_uri, diagnostics, Some(document.version()))
-            .await;
-
-        tracing::debug!(
-            "Published {} diagnostics for {}",
-            diagnostic_count,
-            lsp_uri_text
-        );
-    }
-}
-
-async fn collect_snapshot_diagnostics(
-    snapshot: SessionSnapshot,
-    file: djls_source::File,
-) -> Option<Vec<ls_types::Diagnostic>> {
-    for attempt in 0..=SNAPSHOT_CANCEL_RETRIES {
-        let snapshot = snapshot.clone();
-        let result = tokio::task::spawn_blocking(move || {
-            salsa::Cancelled::catch(AssertUnwindSafe(|| {
-                djls_ide::collect_diagnostics(snapshot.db(), file)
-            }))
-        })
-        .await
-        .expect("diagnostics snapshot task must not panic");
-
-        match result {
-            Ok(diagnostics) => return diagnostics,
-            Err(cancelled) if attempt < SNAPSHOT_CANCEL_RETRIES => {
-                tracing::debug!(
-                    ?cancelled,
-                    attempt = attempt + 1,
-                    "Refresh diagnostics cancelled; retrying with same snapshot"
-                );
-            }
-            Err(cancelled) => {
-                tracing::debug!(
-                    ?cancelled,
-                    retries = SNAPSHOT_CANCEL_RETRIES,
-                    "Refresh diagnostics cancelled; skipping diagnostics republish"
-                );
-                return None;
-            }
-        }
-    }
-
-    unreachable!("diagnostics retry loop must return")
 }
 
 impl LanguageServer for DjangoLanguageServer {
@@ -529,8 +267,7 @@ impl LanguageServer for DjangoLanguageServer {
         tracing::info!("Server received initialized notification.");
 
         // Refresh project data in the background and initialize the workspace.
-        let epoch = self.next_refresh_epoch();
-        self.submit_project_refresh(epoch, true).await;
+        self.submit_project_refresh(true).await;
     }
 
     async fn shutdown(&self) -> LspResult<()> {
@@ -816,23 +553,16 @@ impl LanguageServer for DjangoLanguageServer {
             })
             .await;
 
-        if !settings_update.env_changed
-            && !settings_update.diagnostics_changed
-            && !settings_update.semantic_changed
+        // Any relevant change submits a refresh: the epoch bump supersedes
+        // in-flight refresh work (so a stale republish cannot overwrite newer
+        // settings), and the new task re-applies and republishes diagnostics
+        // for open documents. Bumping without resubmitting would drop a
+        // superseded refresh's apply on the floor.
+        if settings_update.env_changed
+            || settings_update.diagnostics_changed
+            || settings_update.semantic_changed
         {
-            return;
-        }
-
-        if settings_update.env_changed {
-            let epoch = self.next_refresh_epoch();
-            self.submit_project_refresh(epoch, false).await;
-        } else if settings_update.diagnostics_changed || settings_update.semantic_changed {
-            let _epoch = self.next_refresh_epoch();
-            let documents = self.with_session(Session::open_documents).await;
-
-            for document in documents {
-                self.maybe_push_diagnostics(&document).await;
-            }
+            self.submit_project_refresh(false).await;
         }
     }
 }
