+++ crates/djls-server/src/session.rs
@@ -3,6 +3,10 @@
 //! This module implements the LSP session abstraction that manages project-specific
 //! state and the Salsa database for incremental computation.
 
+use std::sync::Arc;
+use std::sync::atomic::AtomicU64;
+use std::sync::atomic::Ordering;
+
 #[cfg(test)]
 use camino::Utf8Path;
 use camino::Utf8PathBuf;
@@ -12,6 +16,7 @@ use djls_project::Db as ProjectDb;
 use djls_source::Db as SourceDb;
 use djls_source::File;
 use djls_source::Offset;
+use tokio::sync::Mutex;
 use tower_lsp_server::ls_types;
 
 use crate::client::ClientInfo;
@@ -22,6 +27,10 @@ use crate::ext::TextDocumentContentChangeEventExt;
 use crate::ext::TextDocumentItemExt;
 use crate::ext::UriExt;
 use crate::workspace::Workspace;
+
+/// How many times snapshot-based reads retry after Salsa cancellation before
+/// giving up and returning a fallback.
+pub(crate) const SNAPSHOT_CANCEL_RETRIES: usize = 2;
 
 /// LSP Session managing project-specific state and database operations.
 ///
@@ -44,6 +53,16 @@ pub(crate) struct Session {
 
     /// The Salsa database for incremental computation
     db: DjangoDatabase,
+
+    /// Generation counter for project refreshes. Background refresh tasks
+    /// capture the value at submission and drop their work when it no longer
+    /// matches — a later bump supersedes any in-flight refresh.
+    refresh_epoch: Arc<AtomicU64>,
+
+    /// Serializes `publishDiagnostics` sends so a refresh-originated republish
+    /// can re-check the epoch under the lock and never overwrite a newer
+    /// publish.
+    diagnostic_publish_lock: Arc<Mutex<()>>,
 }
 
 impl Session {
@@ -83,11 +102,27 @@ impl Session {
             workspace,
             client_info,
             db,
+            refresh_epoch: Arc::new(AtomicU64::new(0)),
+            diagnostic_publish_lock: Arc::new(Mutex::new(())),
         }
     }
 
     pub(crate) fn snapshot(&self) -> SessionSnapshot {
         SessionSnapshot::new(self.db.clone(), self.client_info.clone())
+    }
+
+    /// Advance the refresh epoch, superseding any in-flight refresh task.
+    /// Returns the new epoch for the work that follows the bump.
+    pub(crate) fn bump_refresh_epoch(&self) -> u64 {
+        self.refresh_epoch.fetch_add(1, Ordering::AcqRel) + 1
+    }
+
+    pub(crate) fn refresh_epoch(&self) -> Arc<AtomicU64> {
+        Arc::clone(&self.refresh_epoch)
+    }
+
+    pub(crate) fn diagnostic_publish_lock(&self) -> Arc<Mutex<()>> {
+        Arc::clone(&self.diagnostic_publish_lock)
     }
 
     pub(crate) fn client_info(&self) -> &ClientInfo {
